requirements.txt
asgiref==3.8.1
Django==5.1.6
django-browser-reload==1.18.0
django-debug-toolbar==5.0.1
sqlparse==0.5.3
mysqlclient==2.2.7
pillow==11.1.0
python-dotenv==1.0.1
tornado==6.4.2
tzdata==2025.1
google-cloud-translate==3.12.0
django-parler==2.3
python-dotenv==1.0.1